Fix a 32bit int overflow in case the total profit goes over 2.147b

+++ include/FlipUtils.hpp
@@ -17,7 +17,7 @@ namespace FlipUtils
 	int JsonMaxInt(const std::vector<nlohmann::json>& data, const std::string& key);
 
 	/* 1000 -> 1k, 1000000 -> 1m etc. */
-	std::string RoundBigNumbers(const int number);
+	std::string RoundBigNumbers(const long number);
 	std::string Round(const double value, const int decimals); /* Round a value with given accuracy */
 	void PrintTitle(const std::string& text); /* #### Prints like this #### */
 	int Clamp(const int value, const int min, const int max);
